fix: Moves globalVars.js to /helpers
