+++ .env.example
@@ -1,2 +1,3 @@
 DB_HOST=
 PORT=
+JWT_SECRET=
