+++ helpers/index.js
@@ -1,1 +1,2 @@
 export { default as HttpError } from './HttpError.js';
+export { default as $EMAIL_REG_EXP } from './globalVars.js';
